[EA Forum only] digest ad in the post page footer
