+++ packages/lesswrong/lib/components.ts
@@ -36,6 +36,7 @@ importComponent("PostsAnalyticsPage", () => require('../components/analytics/Pos
 importComponent("AnalyticsDisclaimers", () => require('../components/analytics/AnalyticsDisclaimers'));
 importComponent("DateRangeModal", () => require('../components/analytics/DateRangeModal'));
 
+importComponent("DigestAd", () => require('../components/ea-forum/DigestAd'));
 importComponent("EAHome", () => require('../components/ea-forum/EAHome'));
 importComponent("EAHomeMainContent", () => require('../components/ea-forum/EAHomeMainContent'));
 importComponent("EAHomeCommunityPosts", () => require('../components/ea-forum/EAHomeCommunityPosts'));
